% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop_list.R
\name{add_relop}
\alias{add_relop}
\title{Add a relop to the end of a relop_list.}
\usage{
add_relop(collector, ops)
}
\arguments{
\item{collector}{a rquery::relop_list}

\item{ops}{a rquery::relop}
}
\value{
a rquery::relop_table_source representing ops's future materialization.
}
\description{
Add a relop to the end of a relop_list.
}
