% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop_list.R
\docType{methods}
\name{apply_right_S4,ANY,relop_list-method}
\alias{apply_right_S4,ANY,relop_list-method}
\title{Add a relop to the end of a relop_list (piped version).}
\usage{
\S4method{apply_right_S4}{ANY,relop_list}(pipe_left_arg, pipe_right_arg,
  pipe_environment, left_arg_name, pipe_string, right_arg_name)
}
\arguments{
\item{pipe_left_arg}{relop operation tree}

\item{pipe_right_arg}{relop_list}

\item{pipe_environment}{environment to evaluate in.}

\item{left_arg_name}{name, if not NULL name of left argument.}

\item{pipe_string}{character, name of pipe operator.}

\item{right_arg_name}{name, if not NULL name of right argument.}
}
\value{
result
}
\description{
Add a relop to the end of a relop_list in pipe notation with collector on the right.
}
