% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_data_frame.R
\name{apply_right_S4,ANY,rquery_db_info-method}
\alias{apply_right_S4,ANY,rquery_db_info-method}
\title{Apply pipeline to a database.}
\usage{
\S4method{apply_right_S4}{ANY,rquery_db_info}(
  pipe_left_arg,
  pipe_right_arg,
  pipe_environment,
  left_arg_name,
  pipe_string,
  right_arg_name
)
}
\arguments{
\item{pipe_left_arg}{relop operation tree}

\item{pipe_right_arg}{rquery_db_info}

\item{pipe_environment}{environment to evaluate in.}

\item{left_arg_name}{name, if not NULL name of left argument.}

\item{pipe_string}{character, name of pipe operator.}

\item{right_arg_name}{name, if not NULL name of right argument.}
}
\value{
result
}
\description{
Apply pipeline to a database with relop %.>% db notation.
}
