% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\docType{methods}
\name{apply_right_S4,data.frame,relop_arrow-method}
\alias{apply_right_S4,data.frame,relop_arrow-method}
\title{S4 dispatch method for apply_right.}
\usage{
\S4method{apply_right_S4}{data.frame,relop_arrow}(pipe_left_arg,
  pipe_right_arg, pipe_environment, left_arg_name, pipe_string,
  right_arg_name)
}
\arguments{
\item{pipe_left_arg}{left argument}

\item{pipe_right_arg}{pipe_right_arg argument}

\item{pipe_environment}{environment to evaluate in}

\item{left_arg_name}{name, if not NULL name of left argument.}

\item{pipe_string}{character, name of pipe operator.}

\item{right_arg_name}{name, if not NULL name of right argument.}
}
\value{
result
}
\description{
compose a data.frame and a relop_arrow class
}
