% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop_list.R
\docType{methods}
\name{apply_right_S4,relop_list,rquery_db_info-method}
\alias{apply_right_S4,relop_list,rquery_db_info-method}
\title{Materialize a stages list in pipe notation with relop_list on the left.}
\usage{
\S4method{apply_right_S4}{relop_list,rquery_db_info}(pipe_left_arg,
  pipe_right_arg, pipe_environment, left_arg_name, pipe_string,
  right_arg_name)
}
\arguments{
\item{pipe_left_arg}{relop_list}

\item{pipe_right_arg}{rquery_db_info}

\item{pipe_environment}{environment to evaluate in.}

\item{left_arg_name}{name, if not NULL name of left argument.}

\item{pipe_string}{character, name of pipe operator.}

\item{right_arg_name}{name, if not NULL name of right argument.}
}
\value{
result
}
\description{
Materialize a stages list in pipe notation with relop_list on the left.
}
