% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{arrow}
\alias{arrow}
\title{Data arrow}
\usage{
arrow(pipeline)
}
\arguments{
\item{pipeline}{pipeline with one source table}
}
\value{
relop_arrow wrapping of pipeline
}
\description{
A categorical arrow mapping a table to a table.
}
