% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize.R
\name{commencify}
\alias{commencify}
\title{Hyderdrive (science fiction show) synonym for \code{\link{execute}}}
\usage{
commencify(source, optree, ..., limit = NULL, source_limit = NULL,
  overwrite = TRUE, temporary = TRUE, precheck = FALSE,
  allow_executor = TRUE, temp_source = mk_tmp_name_source("rquery_ex"),
  env = parent.frame())
}
\arguments{
\item{source}{data.frame or DBI connection.}

\item{optree}{relop operation tree.}

\item{...}{force later arguments to bind by name.}

\item{limit}{numeric, if set limit to this many rows during data bring back (not used when landing a table).}

\item{source_limit}{numeric if not NULL limit sources to this many rows.}

\item{overwrite}{logical if TRUE drop an previous table.}

\item{temporary}{logical if TRUE try to create a temporary table.}

\item{precheck}{logical if TRUE precheck existance of table and columns.}

\item{allow_executor}{logical if TRUE allow any executor set as rquery.rquery_executor to be used.}

\item{temp_source}{temporary name generator.}

\item{env}{environment to work in.}
}
\value{
data.frame
}
\description{
Hyderdrive (science fiction show) synonym for \code{\link{execute}}
}
\seealso{
\code{\link{execute}}
}
