% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect}
\usage{
connect()
}
\value{
ODBC connection object
}
\description{
Connect to the database. Uses the db_dsn, db_username and db_password environment variables. Use Sys.setenv() to set them.
}
\examples{

sys.setenv(
  'db_dsn' = 'mydb',
  'db_username' = 'myuser',
  'db_password' = 'password'
)

conn <- connect()
}
