% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{convert_yaml_to_pipeline}
\alias{convert_yaml_to_pipeline}
\title{Convert a series of simple objects (from YAML deserializaton) to an rquery pipeline.}
\usage{
convert_yaml_to_pipeline(rep, source = NULL)
}
\arguments{
\item{rep}{input objects}

\item{source}{input rquery node}
}
\value{
rquery operator tree
}
\description{
Convert a series of simple objects (from YAML deserializaton) to an rquery pipeline.
}
