% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.directories.R
\name{create.directories}
\alias{create.directories}
\title{Create directories}
\usage{
create.directories(data.dir = "data", sql.dir = "sql",
  subs.dir = "params")
}
\arguments{
\item{data.dir}{Query results directory. Stores the results of the SQL queries as RDS files.}

\item{sql.dir}{SQL query directory. Stores all the SQL used. If subs were provided, it saves the query before the subs were inserted.}

\item{subs.dir}{Subs directory. Stores all the individual subs for each query as a data frame.}
}
\description{
Create the directories required for storing the query data and metadata locally.
}
