% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dbi_fns.R
\name{dbi_colnames}
\alias{dbi_colnames}
\title{Old function name for \code{\link{rq_colnames}} (old alias will eventually be removed).}
\usage{
dbi_colnames(db, table_name, ..., qualifiers = NULL)
}
\arguments{
\item{db}{Connection handle}

\item{table_name}{character table name}

\item{...}{not used, force later argument to bind by name}

\item{qualifiers}{optional named ordered vector of strings carrying additional db hierarchy terms, such as schema.}
}
\value{
character list of column names
}
\description{
Old function name for \code{\link{rq_colnames}} (old alias will eventually be removed).
}
\seealso{
\code{\link{rq_colnames}}
}
