% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dbi_fns.R
\name{dbi_coltypes}
\alias{dbi_coltypes}
\title{Old function name for \code{\link{rq_coltypes}} (old alias will eventually be removed).}
\usage{
dbi_coltypes(db, table_name, ..., prefer_not_NA = FALSE,
  force_check = FALSE)
}
\arguments{
\item{db}{Connection handle.}

\item{table_name}{character table name refering to a non-empty table.}

\item{...}{force later arguments to bind by name.}

\item{prefer_not_NA}{logical, if TRUE try to find an non-NA example for all columns (FALSE just for logical columns).}

\item{force_check}{logical, if TRUE perform checks regardless of check_logical_column_types option setting.}
}
\value{
single row data.frame with example values, not all values necessarily from same database row.
}
\description{
Old function name for \code{\link{rq_coltypes}} (old alias will eventually be removed).
}
\seealso{
\code{\link{rq_coltypes}}
}
