% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dbi_fns.R
\name{dbi_connection_advice}
\alias{dbi_connection_advice}
\title{Old function name for \code{\link{rq_connection_advice}} (old alias will eventually be removed).}
\usage{
dbi_connection_advice(db)
}
\arguments{
\item{db}{database connection handle}
}
\value{
named list of options
}
\description{
Old function name for \code{\link{rq_connection_advice}} (old alias will eventually be removed).
}
\seealso{
\code{\link{rq_connection_advice}}
}
