% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_ops.R
\name{dbi_connection_name}
\alias{dbi_connection_name}
\title{Build a cannonical name for a db connection class.}
\usage{
dbi_connection_name(db)
}
\arguments{
\item{db}{DBI database connection}
}
\value{
character
}
\description{
Build a cannonical name for a db connection class.
}
\examples{

if(requireNamespace("RSQLite", quietly = TRUE)) {
  my_db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
  print(dbi_connection_name(my_db))
  print(dbi_connection_preferences(my_db))
  DBI::dbDisconnect(my_db)
}

}
