% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dbi_fns.R
\name{dbi_connection_tests}
\alias{dbi_connection_tests}
\title{Old function name for \code{\link{rq_connection_tests}} (old alias will eventually be removed).}
\usage{
dbi_connection_tests(db, ..., overrides = NULL, use_advice = TRUE)
}
\arguments{
\item{db}{database connection handle.}

\item{...}{force later arguments to bind by name.}

\item{overrides}{named character vector or list, options (just name, not DB qualification) to force}

\item{use_advice}{logical if TRUE incorpeate hard-coded advice.}
}
\value{
named list of options
}
\description{
Old function name for \code{\link{rq_connection_tests}} (old alias will eventually be removed).
}
\seealso{
\code{\link{rq_connection_tests}}
}
