% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dbi_fns.R
\name{dbi_copy_to}
\alias{dbi_copy_to}
\title{Old function name for \code{\link{rq_copy_to}} (old alias will eventually be removed).}
\usage{
dbi_copy_to(db, table_name, d, ..., overwrite = FALSE, temporary = TRUE,
  rowidcolumn = NULL)
}
\arguments{
\item{db}{database connection handle.}

\item{table_name}{name of table to create.}

\item{d}{data.frame to copy to database.}

\item{...}{force later argument to be by name}

\item{overwrite}{logical, if TRUE try to overwrite existing table.}

\item{temporary}{logical, if TRUE try to mark table as temporary.}

\item{rowidcolumn}{character, name to land row-ids.}
}
\value{
a relop representation of the data
}
\description{
Old function name for \code{\link{rq_copy_to}} (old alias will eventually be removed).
}
\seealso{
\code{\link{rq_copy_to}}
}
