% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_ops.R
\name{dbi_execute}
\alias{dbi_execute}
\title{Execute a query.}
\usage{
dbi_execute(db, q)
}
\arguments{
\item{db}{DBI database connection}

\item{q}{character query}
}
\value{
nothing
}
\description{
Execute a query.
}
\seealso{
\code{\link{dbi_table}}
}
