% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dbi_fns.R
\name{dbi_nrow}
\alias{dbi_nrow}
\title{Old function name for \code{\link{rq_nrow}} (old alias will eventually be removed).}
\usage{
dbi_nrow(db, table_name)
}
\arguments{
\item{db}{database connetion}

\item{table_name}{character, name of table}
}
\value{
numeric row count
}
\description{
Old function name for \code{\link{rq_nrow}} (old alias will eventually be removed).
}
\seealso{
\code{\link{rq_nrow}}
}
