% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dbi_fns.R
\name{dbi_table_exists}
\alias{dbi_table_exists}
\title{Old function name for \code{\link{rq_table_exists}} (old alias will eventually be removed).}
\usage{
dbi_table_exists(db, table_name)
}
\arguments{
\item{db}{Connection handle}

\item{table_name}{character table name}
}
\value{
logical TRUE if table exists.
}
\description{
Old function name for \code{\link{rq_table_exists}} (old alias will eventually be removed).
}
\seealso{
\code{\link{rq_table_exists}}
}
