% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_clutch.R
\name{doubleapply}
\alias{doubleapply}
\alias{\%>>\%}
\title{Placeholder for old double apply, do not use.}
\usage{
pipe_left_arg \%>>\% pipe_right_arg
}
\arguments{
\item{pipe_left_arg}{pipe_left_arg}

\item{pipe_right_arg}{pipe_right_arg}
}
\value{
\code{d \%.>\% ops}
}
\description{
Placeholder for old double apply, do not use.
}
