% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_ex.R
\name{ex}
\alias{ex}
\title{Execute a wrapped execution pipeline.}
\usage{
ex(ops, ..., env = parent.frame())
}
\arguments{
\item{ops}{rquery pipeline with tables formed by `wrap()`.}

\item{...}{not used, force later argument to be referred by name}

\item{env}{environment to work in.}
}
\value{
data.frame result
}
\description{
Execute a ops-dag using `wrap()` data as values.
}
\examples{

if(requireNamespace('rqdatatable')) {
 d <- data.frame(x = 1:3, y = 4:6)
 d \%.>\%
   wrap(.) \%.>\%
   extend(., z := x + y) \%.>\%
   ex(.)
}

}
