% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend.R
\name{extend_nse}
\alias{extend_nse}
\title{Extend data by adding more columns.}
\usage{
extend_nse(source, ..., partitionby = NULL, orderby = NULL,
  reverse = NULL, display_form = NULL, env = parent.frame())
}
\arguments{
\item{source}{source to select from.}

\item{...}{new column assignment expressions.}

\item{partitionby}{partitioning (window function) terms.}

\item{orderby}{ordering (in window function) terms.}

\item{reverse}{reverse ordering (in window function) terms.}

\item{display_form}{chacter presentation form}

\item{env}{environment to look for values in.}
}
\value{
extend node.
}
\description{
Create a node similar to a Codd extend relational operator (add derived columns).
}
\details{
Partitionby and orderby can only be used with a database that supports window-functions
(such as PostgreSQL, Spark, and so on).
}
\examples{

if (requireNamespace("DBI", quietly = TRUE) && requireNamespace("RSQLite", quietly = TRUE)) {
  my_db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
  d <- rq_copy_to(my_db, 'd',
                   data.frame(AUC = 0.6, R2 = 0.2))
  optree <- extend_nse(d, v \%:=\% ifelse(AUC>0.5, R2, 1.0))
  cat(format(optree))
  sql <- to_sql(optree, my_db)
  cat(sql)
  print(DBI::dbGetQuery(my_db, sql))
  DBI::dbDisconnect(my_db)
}

}
