% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop.R
\name{format_node}
\alias{format_node}
\title{Format a single node for printing.}
\usage{
format_node(node)
}
\arguments{
\item{node}{node of operator tree to be formatted}
}
\value{
character disply form of the node
}
\description{
Format a single node for printing.
}
