% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.suffix.R
\name{get.suffix}
\alias{get.suffix}
\title{Get suffix}
\usage{
get.suffix(subs)
}
\arguments{
\item{subs}{One row data frame whose column names are the substrings to replace, and whose values are the strings to replace them with.}
}
\description{
Get the file suffix as a hash of the subs list if it is provided. Required to keep uniqueness of various queries.
}
