% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_controller.R
\name{key_inspector_postgresql}
\alias{key_inspector_postgresql}
\title{Return all primary key columns as guess at preferred primary keys for a PostgreSQL handle.}
\usage{
key_inspector_postgresql(db, tablename)
}
\arguments{
\item{db}{database handle}

\item{tablename}{character, name of table}
}
\value{
map of keys to keys
}
\description{
Return all primary key columns as guess at preferred primary keys for a PostgreSQL handle.
}
\seealso{
\code{describe_tables}
}
