% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.data.R
\name{load.data}
\alias{load.data}
\title{Load data}
\usage{
load.data(name, subs = NULL, data.dir = "data")
}
\arguments{
\item{name}{Name of the query}

\item{data.dir}{Query results directory. Stores the results of the SQL queries as RDS files.}
}
\value{
Data frame of the query results
}
\description{
Load the query results from the filesystem.
}
