% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.data.R, R/load.sql.R
\name{load.sql}
\alias{load.sql}
\alias{load.sql}
\title{Load data}
\usage{
load.sql(name, sql.dir = "sql")

load.sql(name, sql.dir = "sql")
}
\arguments{
\item{name}{Name of the query}

\item{sql.dir}{SQL query directory. Stores all the SQL used. If subs were provided, it saves the query before the subs were inserted.}

\item{data.dir}{Query results directory. Stores the results of the SQL queries as RDS files.}

\item{name}{Name of the query}
}
\value{
Data frame of the query results

SQL used for the query
}
\description{
Load the query results from the filesystem.

Load the subs list from the filesystem.
}
