% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.subs.R
\name{load.subs}
\alias{load.subs}
\title{Load subs}
\usage{
load.subs(name, subs.dir = "subs")
}
\arguments{
\item{name}{Name of the query}

\item{subs.dir}{Subs directory. Stores all the individual subs for each query as a data frame.}
}
\value{
Data frame of all subs used for this query, or NULL if none used
}
\description{
Load the subs list from the filesystem.
}
