% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop_list.R
\name{make_relop_list}
\alias{make_relop_list}
\title{Create a new \code{rquery::relop} operator tree collector list}
\usage{
make_relop_list(name_source = wrapr::mk_tmp_name_source("rqrol"))
}
\arguments{
\item{name_source}{a wrapr::mk_tmp_name_source()}
}
\value{
a relop_list relop stage collector
}
\description{
Create a new \code{rquery::relop} operator tree collector list
}
