% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop_list.R
\name{materialize_relop_list_stages}
\alias{materialize_relop_list_stages}
\title{Materialize the stages list.}
\usage{
materialize_relop_list_stages(db, collector, ..., limit = NULL,
  source_limit = NULL, overwrite = TRUE, temporary = FALSE)
}
\arguments{
\item{db}{database connecton (rquery_db_info class or DBI connections preferred).}

\item{collector}{a rquery::relop_list}

\item{...}{force later arguments to bind by name.}

\item{limit}{numeric if not NULL result limit (to use this, last statement must not have a limit).}

\item{source_limit}{numeric if not NULL limit sources to this many rows.}

\item{overwrite}{logical if TRUE drop an previous table.}

\item{temporary}{logical if TRUE try to create a temporary table.}
}
\value{
a rquery::relop_table_source representing ops's materialization.
}
\description{
Materialize the stages list.
}
