% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora.connect.string.R
\name{ora.connect.string}
\alias{ora.connect.string}
\title{Oracle connection string}
\usage{
ora.connect.string()
}
\value{
Nothing. Sets the environment variables.
}
\description{
Generate the connection string environment variable for Oracle environments only. The host, port and sid should be set using Sys.setenv(db_host, db_port, db_sid) respectively.
}
\examples{

Sys.setenv(db_host = 'myhost', db_port = 123, db_sid = 'mysid')

# Sets the "connect.string" environment variable
ora.connect.string()

}
