% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_sql.R
\name{pre_sql_fn}
\alias{pre_sql_fn}
\title{pre_sql_token funtion name}
\usage{
pre_sql_fn(value)
}
\arguments{
\item{value}{function tname}

\item{value}{character, token string}
}
\value{
pre_sql_token

pre_sql_token
}
\description{
pre_sql_token funtion name
}
\keyword{internal}
