% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_sql.R
\name{pre_sql_identifier}
\alias{pre_sql_identifier}
\title{pre_sql_identifier: abstract name of a column and where it is comming from}
\usage{
pre_sql_identifier(column_name)
}
\arguments{
\item{column_name}{character name of column}
}
\value{
pre_sql_token
}
\description{
represents a value from a given sub-table or sub-expression
 source is name of source
 name is name for term
}
\keyword{internal}
