% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_sql.R
\name{pre_sql_string}
\alias{pre_sql_string}
\title{pre_sql_string}
\usage{
pre_sql_string(value)
}
\arguments{
\item{value}{string}
}
\value{
pre_sql_token
}
\description{
represents a string constant
  value character string
}
\keyword{internal}
