% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_sql.R
\name{pre_sql_sub_expr}
\alias{pre_sql_sub_expr}
\title{pre_sql_sub_expr}
\usage{
pre_sql_sub_expr(terms, info = NULL)
}
\arguments{
\item{terms}{list of pre_sql tokens}

\item{info}{named list of extra info with a name slot containing a single string without spaces.}
}
\value{
pre_sql_sub_expr
}
\description{
represents an expression.  Unnamed list of pre_sql_terms and character.
}
