% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_sql.R
\name{pre_sql_to_query.pre_sql_sub_expr}
\alias{pre_sql_to_query.pre_sql_sub_expr}
\title{Convert a pre_sql token object to SQL query text.}
\usage{
\method{pre_sql_to_query}{pre_sql_sub_expr}(x, db_info, ...,
  source_table = NULL, source_limit = NA_real_, using = NULL,
  qualifiers = NULL)
}
\arguments{
\item{x}{the pre_sql token}

\item{db_info}{representation of the database to convert to}

\item{...}{force later arguments to be by name}

\item{source_table}{concrete table for query}

\item{source_limit}{numeric limit on rows from this source table}

\item{using}{TBD}

\item{qualifiers}{named ordered vector of strings carrying additional db hierarchy terms, such as schema.}
}
\value{
SQL query text
}
\description{
Convert a pre_sql token object to SQL query text.
}
\keyword{internal}
