% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_sql.R
\name{pre_sql_to_query}
\alias{pre_sql_to_query}
\title{Return SQL transform of tokens.}
\usage{
pre_sql_to_query(x, db_info, ..., source_table = NULL,
  source_limit = NA_real_, using = NULL, qualifiers = NULL)
}
\arguments{
\item{x}{parsed tokens.}

\item{db_info}{DBI connnection or rquery_db_info object}

\item{...}{generic additional arguments (not used).}

\item{source_table}{character if not NULL name of source table.}

\item{source_limit}{numeric if not NULL limit sources to this many rows.}

\item{using}{character, if not NULL set of columns used from above.}

\item{qualifiers}{named ordered vector of strings carrying additional db hierarchy terms, such as schema.}
}
\value{
SQL command
}
\description{
Return SQL transform of tokens.
}
\keyword{internal}
