% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_sql.R
\name{pre_sql_token}
\alias{pre_sql_token}
\title{pre_sql_token}
\usage{
pre_sql_token(value)
}
\arguments{
\item{value}{character, token string}
}
\value{
pre_sql_token
}
\description{
general token
}
\keyword{internal}
