% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Query}
\usage{
query(sql, subs = NULL)
}
\arguments{
\item{sql}{SQL query to run}

\item{subs}{One row data frame whose column names are the substrings to replace, and whose values are the strings to replace them with.}
}
\description{
Query the database
}
\examples{

sys.setenv(
  'db_dsn' = 'mydb',
  'db_username' = 'myuser',
  'db_password' = 'password'
)

sql <- "select 1 as a from dual where 1 = _number_"
subs <- data.frame('_number_' = 1)

query(sql, subs)
}
