% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_info.R
\name{quote_identifier}
\alias{quote_identifier}
\title{Quote an identifier.}
\usage{
quote_identifier(x, id)
}
\arguments{
\item{x}{database handle or rquery_db_info object.}

\item{id}{character to quote}
}
\value{
quoted identifier
}
\description{
Quote an identifier.
}
