% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_info.R
\name{quote_literal}
\alias{quote_literal}
\title{Quote a value}
\usage{
quote_literal(x, o)
}
\arguments{
\item{x}{database handle or rquery_db_info object.}

\item{o}{value to quote}
}
\value{
quoted string
}
\description{
Quote a value
}
