% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_info.R
\name{quote_string}
\alias{quote_string}
\title{Quote a string}
\usage{
quote_string(x, s)
}
\arguments{
\item{x}{database handle or rquery_db_info object.}

\item{s}{character to quote}
}
\value{
quoted string
}
\description{
Quote a string
}
