% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_info.R
\name{quote_table_name}
\alias{quote_table_name}
\title{Quote a table name.}
\usage{
quote_table_name(x, id, ..., qualifiers = character(0))
}
\arguments{
\item{x}{database handle or rquery_db_info object.}

\item{id}{character to quote}

\item{...}{not used, force later arguments to bind by name.}

\item{qualifiers}{named ordered vector of strings carrying additional db hierarchy terms, such as schema.}
}
\value{
quoted identifier
}
\description{
Quote a table name.
}
