% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{refresh}
\alias{refresh}
\title{Refresh data}
\usage{
refresh(daily = TRUE, overwrite = FALSE, data.dir = "data",
  sql.dir = "sql", subs.dir = "subs")
}
\arguments{
\item{daily}{Boolean to indicate whether to force a rerun of the query on first execution each day.}

\item{overwrite}{Boolean to indicate whether to force a rerun of the query.}

\item{data.dir}{Query results directory. Stores the results of the SQL queries as RDS files.}

\item{sql.dir}{SQL query directory. Stores all the SQL used. If subs were provided, it saves the query before the subs were inserted.}

\item{subs.dir}{Subs directory. Stores all the individual subs for each query as a data frame.}

\item{sql}{SQL query to run}

\item{name}{name of query}

\item{subs}{One row data frame whose column names are the substrings to replace, and whose values are the strings to replace them with.}
}
\value{
Data frame of the query results
}
\description{
Refresh all applicable data files.
}
\examples{

# Run a daily refresh
refresh()

# Force a refresh
refresh(overwrite = TRUE)
}

