% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop_list.R
\docType{class}
\name{relop_list-class}
\alias{relop_list-class}
\title{List of \code{rquery::relop} operator trees taken in order.}
\description{
Used to collect a sequence of related operations.
For how to use please see here \url{https://github.com/WinVector/rquery/blob/master/extras/query_growth/query_growth.md}.
}
\seealso{
\code{\link{add_relop}}, \code{\link{get_relop_list_stages}}, \code{\link{materialize_relop_list_stages}}
}
