% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_ops.R
\name{rq_execute}
\alias{rq_execute}
\title{Execute a query, typcially an update that is not supposed to return results.}
\usage{
rq_execute(db, q)
}
\arguments{
\item{db}{database connection handle}

\item{q}{character query}
}
\value{
nothing
}
\description{
Execute a query, typcially an update that is not supposed to return results.
}
\seealso{
\code{\link{db_td}}
}
