% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_info.R
\name{rq_function_mappings}
\alias{rq_function_mappings}
\title{Return function mappings for a connection}
\usage{
rq_function_mappings(db, expr_map = NULL)
}
\arguments{
\item{db}{a rquery_db_info}

\item{expr_map}{rquery expression map to include in report (function names to pseudo pre-sql sequences)}
}
\value{
data.frame of function mappings
}
\description{
Return function mappings for a connection
}
\keyword{internal}
