% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_ops.R
\name{rq_get_query}
\alias{rq_get_query}
\title{Execute a get query, typcially a non-update that is supposed to return results.}
\usage{
rq_get_query(db, q)
}
\arguments{
\item{db}{database connection handle}

\item{q}{character query}
}
\value{
nothing
}
\description{
Execute a get query, typcially a non-update that is supposed to return results.
}
\seealso{
\code{\link{db_td}}
}
