% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_ops.R
\name{rq_head}
\alias{rq_head}
\title{Get head of db table}
\usage{
rq_head(db, table_name, ..., qualifiers = NULL, limit = 6L)
}
\arguments{
\item{db}{Connection handle}

\item{table_name}{character table name}

\item{...}{not used, force later argument to bind by name}

\item{qualifiers}{optional named ordered vector of strings carrying additional db hierarchy terms, such as schema.}

\item{limit}{integer, how many rows to take}
}
\value{
first few rows
}
\description{
Get head of db table
}
