\name{rQuery.add}
\alias{rQuery.add}
\title{Add one or more element to an rQueryResult container.}
\arguments{
  \item{rquery}{rQueryResult container}

  \item{value}{\code{XML} element or rQueryResult to add to
  \code{rquery}}
}
\value{
  a new \code{rQueryResult} containing the combined
  elements.
}
\description{
  Add one or more element to an rQueryResult container.
}
\author{
  jfolson
}

