\name{rQuery.addClass}
\alias{rQuery.addClass}
\title{Adds the specified class(es) to each of the set of matched elements.}
\arguments{
  \item{rquery}{rQueryResult container}

  \item{class}{a new class value}
}
\value{
  rQueryResult with the modified elements
}
\description{
  Adds the specified class(es) to each of the set of
  matched elements.
}
\author{
  jfolson
}

