\name{rQuery.after}
\alias{rQuery.after}
\title{Insert content, specified by the parameter, after each element in the set of
matched elements.}
\arguments{
  \item{rquery}{rQueryResult container}

  \item{value}{\code{XML} element, \code{rQueryResult}
  container, or character vector of html}
}
\value{
  the modified \code{rquery} object
}
\description{
  Insert content, specified by the parameter, after each
  element in the set of matched elements.
}
\author{
  jfolson
}

