\name{rQuery.attr}
\alias{rQuery.attr}
\title{Get the value of an attribute for the first element in the set of matched
elements or set one or more attributes for every matched element.}
\arguments{
  \item{rquery}{rQueryResult container}

  \item{...}{a namded list of attribute values}
}
\value{
  list of attribute values or a rQueryResult with the
  modified elements
}
\description{
  Get the value of an attribute for the first element in
  the set of matched elements or set one or more attributes
  for every matched element.
}
\author{
  jfolson
}

