\name{rQuery.class}
\alias{rQuery.class}
\title{Get the class value(s) for the first element in the set of matched
elements or add a class for every matched element.}
\arguments{
  \item{rquery}{rQueryResult container}

  \item{class}{a new class value}
}
\value{
  list of class values or a rQueryResult with the modified
  elements
}
\description{
  Get the class value(s) for the first element in the set
  of matched elements or add a class for every matched
  element.
}
\author{
  jfolson
}

