\name{rQuery.find}
\alias{rQuery.find}
\title{Get the descendants of each element in the current set of matched elements,
filtered by a selector, jQuery object, or element.}
\arguments{
  \item{rquery}{rQueryResult container to search}

  \item{query}{CSS selector query}
}
\value{
  XMLAbstractNode wrapped in a rQueryResult
}
\description{
  Get the descendants of each element in the current set of
  matched elements, filtered by a selector, jQuery object,
  or element.
}
\author{
  jfolson
}

