\name{rQuery.hasClass}
\alias{rQuery.hasClass}
\title{Determine whether any of the matched elements are assigned the given class.}
\arguments{
  \item{rquery}{rQueryResult container}

  \item{class}{a new class value}
}
\value{
  vector of booleans
}
\description{
  Determine whether any of the matched elements are
  assigned the given class.
}
\author{
  jfolson
}

