\name{rQuery.html}
\alias{rQuery.html}
\title{Get the HTML contents of the first element in the set of matched elements or set the HTML contents of every matched element.}
\arguments{
  \item{rquery}{an rQueryResult object}

  \item{value}{optional \code{XML} object or html character
  vector.}
}
\value{
  rQueryResult containing the modified \code{rquery}
  element or a character vector containing the html
  contents of the first element.
}
\description{
  Get the HTML contents of the first element in the set of
  matched elements or set the HTML contents of every
  matched element.
}

