\name{rQuery.insertAfter}
\alias{rQuery.insertAfter}
\title{Insert every element in the set of matched elements after the target.}
\arguments{
  \item{rquery}{rQueryResult container}

  \item{value}{\code{XML} element, \code{rQueryResult}
  container, or character vector of html}
}
\value{
  the modified \code{rquery} object
}
\description{
  Insert every element in the set of matched elements after
  the target.
}
\author{
  jfolson
}

