\name{rQuery.prepend}
\alias{rQuery.prepend}
\title{Insert content, specified by the parameter, to the beginning of each element in the
set of matched elements.}
\arguments{
  \item{rquery}{rQueryResult container}

  \item{value}{\code{XML} element, \code{rQueryResult}
  container, or character vector of html}
}
\value{
  the modified \code{rquery} object
}
\description{
  Insert content, specified by the parameter, to the
  beginning of each element in the set of matched elements.
}
\author{
  jfolson
}

