% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rquery.R
\name{rquery}
\alias{rquery}
\title{RQuery}
\usage{
rquery(sql = NULL, name, subs = NULL, data.dir = "data",
  sql.dir = "sql", subs.dir = "subs", daily = FALSE, overwrite = FALSE)
}
\arguments{
\item{sql}{SQL query to run}

\item{name}{name of query}

\item{subs}{One row data frame whose column names are the substrings to replace, and whose values are the strings to replace them with.}

\item{data.dir}{Query results directory. Stores the results of the SQL queries as RDS files.}

\item{sql.dir}{SQL query directory. Stores all the SQL used. If subs were provided, it saves the query before the subs were inserted.}

\item{subs.dir}{Subs directory. Stores all the individual subs for each query as a data frame.}

\item{daily}{Boolean to indicate whether to force a rerun of the query on first execution each day.}

\item{overwrite}{Boolean to indicate whether to force a rerun of the query.}
}
\value{
Data frame of the query results
}
\description{
Query the database efficiently and cache data appropriately in subdirectories
}
\examples{

sys.setenv(
  'db_dsn' = 'mydb',
  'db_username' = 'myuser',
  'db_password' = 'password'
)

sql <- "select 1 as a from dual where 1 = XnumberX"
subs <- data.frame('XnumberX' = 1)

rquery(sql, 'myquery', subs)
}

