% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_info.R
\name{rquery_db_info}
\alias{rquery_db_info}
\title{Build a db information stand-in}
\usage{
rquery_db_info(indentifier_quote_char, string_quote_char)
}
\arguments{
\item{indentifier_quote_char}{character}

\item{string_quote_char}{character}
}
\value{
rquery_db_info object
}
\description{
Build a db information stand-in
}
