% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_info.R
\docType{data}
\name{rquery_default_db_info}
\alias{rquery_default_db_info}
\title{An example \code{rquery_db_info} object useful for formatting \code{SQL} without a database connection.}
\format{An object of class \code{rquery_db_info} of length 10.}
\usage{
rquery_default_db_info
}
\description{
An example \code{rquery_db_info} object useful for formatting \code{SQL} without a database connection.
}
\keyword{datasets}
