% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/should.refresh.data.R
\name{should.refresh.data}
\alias{should.refresh.data}
\title{Should refresh data}
\usage{
should.refresh.data(name, subs = NULL, data.dir = "data", daily = FALSE,
  overwrite = FALSE)
}
\arguments{
\item{name}{name of query}

\item{subs}{One row data frame whose column names are the substrings to replace, and whose values are the strings to replace them with.}

\item{data.dir}{Query results directory. Stores the results of the SQL queries as RDS files.}

\item{daily}{Boolean to indicate whether to force a rerun of the query on first execution each day.}

\item{overwrite}{Boolean to indicate whether to force a rerun of the query.}
}
\value{
Boolean indicating whether the SQL query should be rerun
}
\description{
Ascertain whether the SQL query should be rerun
}

