% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop_list.R
\docType{methods}
\name{show,relop_list-method}
\alias{show,relop_list-method}
\title{S4 print method}
\usage{
\S4method{show}{relop_list}(object)
}
\arguments{
\item{object}{item to print}
}
\description{
S4 print method
}
