% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.sql.R
\name{store.sql}
\alias{store.sql}
\title{Store SQL}
\usage{
store.sql(name, sql = NULL, sql.dir = "sql")
}
\arguments{
\item{name}{Name of the query}

\item{sql}{SQL query to run}

\item{sql.dir}{SQL query directory. Stores all the SQL used. If subs were provided, it saves the query before the subs were inserted.}
}
\description{
Store the SQL query to the filesystem.
}

