% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.subs.R
\name{store.subs}
\alias{store.subs}
\title{Store subs}
\usage{
store.subs(name, subs = NULL, subs.dir = "subs")
}
\arguments{
\item{name}{Name of the query}

\item{subs}{One row data frame whose column names are the substrings to replace, and whose values are the strings to replace them with.}

\item{subs.dir}{Subs directory. Stores all the individual subs for each query as a data frame.}
}
\description{
Store the subs list to the filesystem.
}

