% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_sql.R
\name{str_pre_sql_sub_expr}
\alias{str_pre_sql_sub_expr}
\title{Structure of a pre_sql_sub_expr}
\usage{
str_pre_sql_sub_expr(x)
}
\arguments{
\item{x}{a pre_sql_sub_expr}
}
\value{
charcter presentation with {} denoting nesting
}
\description{
Structure of a pre_sql_sub_expr
}
\keyword{internal}
