% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop.R
\name{to_pre_sql}
\alias{to_pre_sql}
\title{Return SQL implementation of operation tree.}
\usage{
to_pre_sql(x, ...)
}
\arguments{
\item{x}{rquery operation tree.}

\item{...}{generic additional arguments (not used).}
}
\value{
pre_sql_op node
}
\description{
Return SQL implementation of operation tree.
}
