% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop.R
\name{to_sql}
\alias{to_sql}
\title{Return SQL implementation of operation tree.}
\usage{
to_sql(x, db, ..., source_limit = NULL, indent_level = 0,
  tnum = mkTempNameGenerator("tsql"), append_cr = TRUE, using = NULL)
}
\arguments{
\item{x}{rquery operation tree.}

\item{db}{DBI database handle or rquery_db_info object.}

\item{...}{generic additional arguments (not used).}

\item{source_limit}{numeric if not NULL limit sources to this many rows.}

\item{indent_level}{level to indent.}

\item{tnum}{temp sub-query name generator.}

\item{append_cr}{logical if TRUE end with CR.}

\item{using}{character, if not NULL set of columns used from above.}
}
\value{
SQL command
}
\description{
Return SQL implementation of operation tree.
}
