% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{to_transport_representation}
\alias{to_transport_representation}
\title{Convert an rquery op diagram to a simple representation, appropriate for conversion to YAML.}
\usage{
to_transport_representation(ops)
}
\arguments{
\item{ops}{rquery operator dag}
}
\value{
represenation structure
}
\description{
Convert an rquery op diagram to a simple representation, appropriate for conversion to YAML.
}
