% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_for_sql.R
\name{tokenize_for_SQL}
\alias{tokenize_for_SQL}
\title{Cross-parse from an R parse tree into SQL.}
\usage{
tokenize_for_SQL(lexpr, colnames, env = parent.frame())
}
\arguments{
\item{lexpr}{item from  \code{substitute}}

\item{colnames}{column names of table}

\item{env}{environment to look for values}
}
\value{
sql info: list(presentation, parsed_toks(list of tokens), sql_text, symbols_used, symbols_produced, free_symbols)
}
\description{
Cross-parse from an R parse tree into SQL.
}
\examples{

tokenize_for_SQL(substitute(1 + 2), colnames= NULL)
tokenize_for_SQL(substitute(a := 3), colnames= NULL)
tokenize_for_SQL(substitute(a \%:=\% ( 3 + 4 )), colnames= NULL)
tokenize_for_SQL(substitute(a \%:=\% rank(3, 4)), colnames= NULL)

}
