% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_data_frame.R
\name{wrapr_function.relop}
\alias{wrapr_function.relop}
\title{Execute pipeline treating pipe_left_arg as local data to
be copied into database.}
\usage{
\method{wrapr_function}{relop}(pipe_left_arg, pipe_right_arg, pipe_environment,
  pipe_name = NULL)
}
\arguments{
\item{pipe_left_arg}{data.frame}

\item{pipe_right_arg}{rquery relop operation tree}

\item{pipe_environment}{environment to execute in}

\item{pipe_name}{name of pipling symbol}
}
\value{
data.frame
}
\description{
Execute pipeline treating pipe_left_arg as local data to
be copied into database.
}
\seealso{
\code{\link{rquery_apply_to_data_frame}}
}
