% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop_list.R
\name{add_relop}
\alias{add_relop}
\title{Add a relop to the end of a relop_list.}
\usage{
add_relop(collector, ops)
}
\arguments{
\item{collector}{a rquery::relop_list}

\item{ops}{a rquery::relop}
}
\value{
a rquery::relop_table_source representing ops's future materialization.
}
\description{
For how to use please see here \url{https://github.com/WinVector/rquery/blob/master/extras/query_growth/query_growth.md}.
}
\examples{

collector <- make_relop_list()
ops <- mk_td("d", c("x", "y")) \%.>\%
  extend(., z \%:=\% x + y) \%.>\%
  collector
get_relop_list_stages(collector)

}
\seealso{
\code{\link{make_relop_list}}, \code{\link{get_relop_list_stages}}, \code{\link{materialize_relop_list_stages}}
}
