% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{arrow}
\alias{arrow}
\title{Data arrow}
\usage{
arrow(pipeline, ..., free_table_key = NULL, strict = FALSE)
}
\arguments{
\item{pipeline}{pipeline with one source table}

\item{...}{not used, force later argument to be referred to by name.}

\item{free_table_key}{name of table to consider free (input) to the pipeline}

\item{strict}{logical, if TRUE excess columns are considered an error}
}
\value{
relop_arrow wrapping of pipeline
}
\description{
A categorical arrow mapping a table to a table.
}
