% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop.R
\name{column_names}
\alias{column_names}
\title{Return column names}
\usage{
column_names(x, ...)
}
\arguments{
\item{x}{rquery operation tree.}

\item{...}{generic additional arguments}
}
\value{
vector of column names
}
\description{
Return column names
}
