% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop.R
\name{columns_used}
\alias{columns_used}
\title{Return columns used}
\usage{
columns_used(x, ..., using = NULL, contract = FALSE)
}
\arguments{
\item{x}{rquery operation tree.}

\item{...}{generic additional arguments (not used)}

\item{using}{character, if not NULL set of columns used from above.}

\item{contract}{logical, if TRUE perform unused value elimination.}
}
\value{
vector of table qualified column names.
}
\description{
Return columns used
}
