% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect}
\usage{
connect(oracle = TRUE)
}
\arguments{
\item{oracle}{Boolean to indicate whether to automatically generate the connect.string for use with Oracle. Otherwise, you can provide the connect string by doing Sys.setenv('a connection string')}
}
\value{
ODBC connection object
}
\description{
Connect to the database. Uses the db_dsn, db_username and db_password environment variables. Use Sys.setenv() to set them.
}
\examples{

Sys.setenv(db_driver = 'mydriver',
           db_host = 'myhost',
           db_port = 123,
           db_sid = 'myservicename',
           db_username = 'myuser',
           db_password = 'mypass')

# If using the Oracle method
ora.connect.string
conn <- connect(oracle = TRUE)

# Or if using the connection string method
Sys.setenv(connect.string = 'a connection string',
           db_username = 'myuser', db_password = 'mypass')

conn <- connect()
}
