% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{convert_yaml_to_pipeline}
\alias{convert_yaml_to_pipeline}
\title{Convert a series of simple objects (from YAML deserializaton) to an rquery pipeline.}
\usage{
convert_yaml_to_pipeline(rep, ..., source = NULL, env = parent.frame())
}
\arguments{
\item{rep}{input objects}

\item{...}{not used, force later arguments to bind by name}

\item{source}{input rquery node}

\item{env}{environment to evaluate in}
}
\value{
rquery operator tree
}
\description{
Convert a series of simple objects (from YAML deserializaton) to an rquery pipeline.
}
