% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_ops.R
\name{dbi_colnames}
\alias{dbi_colnames}
\title{List fields from a dbi connection}
\usage{
dbi_colnames(my_db, tableName)
}
\arguments{
\item{my_db}{DBI connection}

\item{tableName}{character table name}
}
\value{
character list of column names
}
\description{
List fields from a dbi connection
}
