% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dbi_fns.R
\name{dbi_connection_name}
\alias{dbi_connection_name}
\title{Old function name for \code{\link{rq_connection_name}} (old alias will eventually be removed).}
\usage{
dbi_connection_name(db)
}
\arguments{
\item{db}{Database connection handle.}
}
\value{
character, key version of handle for option lookups.
}
\description{
Old function name for \code{\link{rq_connection_name}} (old alias will eventually be removed).
}
\seealso{
\code{\link{rq_connection_name}}
}
