% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_ops.R
\name{dbi_nrow}
\alias{dbi_nrow}
\title{Count rows and return as numeric}
\usage{
dbi_nrow(db, table_name)
}
\arguments{
\item{db}{database connetion}

\item{table_name}{character, name of table}
}
\value{
numeric row count
}
\description{
Count rows and return as numeric
}
