% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_dbi_fns.R
\name{dbi_remove_table}
\alias{dbi_remove_table}
\title{Old function name for \code{\link{rq_remove_table}} (old alias will eventually be removed).}
\usage{
dbi_remove_table(db, table_name)
}
\arguments{
\item{db}{database connection.}

\item{table_name}{character, name of table to create.}
}
\value{
logical TRUE if table existed, else FALSE
}
\description{
Old function name for \code{\link{rq_remove_table}} (old alias will eventually be removed).
}
\seealso{
\code{\link{rq_remove_table}}
}
