% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_ops.R
\name{dbi_table_exists}
\alias{dbi_table_exists}
\title{Check if a table exists.}
\usage{
dbi_table_exists(db, table_name)
}
\arguments{
\item{db}{DBI connection}

\item{table_name}{character table name}
}
\value{
logical TRUE if table exists.
}
\description{
Check if a table exists.
}
\seealso{
\code{\link{dbi_table}}
}
