% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_controller.R
\name{example_employee_date}
\alias{example_employee_date}
\title{build some example tables}
\usage{
example_employee_date(con)
}
\arguments{
\item{con}{db connection}
}
\value{
example tables
}
\description{
build some example tables
}
\examples{

if (requireNamespace("DBI", quietly = TRUE) && requireNamespace("RSQLite", quietly = TRUE)) {
  my_db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
  example_employee_date(my_db)
  DBI::dbDisconnect(my_db)
}

}
