% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop_list.R
\name{get_relop_list_stages}
\alias{get_relop_list_stages}
\title{Return the stages list.}
\usage{
get_relop_list_stages(collector)
}
\arguments{
\item{collector}{a rquery::relop_list}
}
\value{
a list of rquery::relops
}
\description{
Return the stages list.
}
