% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.sql.R
\name{load.sql}
\alias{load.sql}
\title{Load subs}
\usage{
load.sql(name, sql.dir = "sql")
}
\arguments{
\item{name}{Name of the query}

\item{sql.dir}{SQL query directory. Stores all the SQL used. If subs were provided, it saves the query before the subs were inserted.}
}
\value{
SQL used for the query
}
\description{
Load the subs list from the filesystem.
}

