% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_sql_node.R
\name{materialize_node}
\alias{materialize_node}
\title{Cache results to a named table inside a pipeline.}
\usage{
materialize_node(source, table_name, ..., temporary = TRUE)
}
\arguments{
\item{source}{source to work from (relop node)}

\item{table_name}{character, name of caching table}

\item{...}{force later arguments to bind by name}

\item{temporary}{logical, if TRUE mark tables temporary.}
}
\value{
sql node.
}
\description{
Cache results to a named table inside a pipeline.
}
\seealso{
\code{\link{rsummary_node}}, \code{\link{non_sql_node}}
}
