% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_ops.R
\name{rq_colnames}
\alias{rq_colnames}
\title{List table column names.}
\usage{
rq_colnames(db, table_name)
}
\arguments{
\item{db}{Connection handle}

\item{table_name}{character table name}
}
\value{
character list of column names
}
\description{
List table column names.
}
