% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_ops.R
\name{rq_nrow}
\alias{rq_nrow}
\title{Count rows and return as numeric}
\usage{
rq_nrow(db, table_name)
}
\arguments{
\item{db}{database connection}

\item{table_name}{character, name of table}
}
\value{
numeric row count
}
\description{
Count rows and return as numeric
}
\seealso{
\code{\link{db_td}}
}
