% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_ops.R
\name{rq_table_exists}
\alias{rq_table_exists}
\title{Check if a table exists.}
\usage{
rq_table_exists(db, table_name)
}
\arguments{
\item{db}{Connection handle}

\item{table_name}{character table name}
}
\value{
logical TRUE if table exists.
}
\description{
Check if a table exists.
}
\seealso{
\code{\link{db_td}}
}
