\name{rQuery}
\alias{\%$\%}
\alias{rQuery}
\title{jQuery for R}
\usage{
  rQuery(elem, query)

  %$%(elem, query)
}
\arguments{
  \item{elem}{html element to query}

  \item{query}{CSS selector query}
}
\value{
  XMLAbstractNode wrapped in a rQueryResult
}
\description{
  jQuery for R
}
\author{
  jfolson

  jfolson
}

