% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_info.R
\name{rquery_default_db_info}
\alias{rquery_default_db_info}
\title{An example \code{rquery_db_info} object useful for formatting \code{SQL} without a database connection.}
\usage{
rquery_default_db_info()
}
\value{
a rquery_db_info without a connection and vanilla settings.
}
\description{
An example \code{rquery_db_info} object useful for formatting \code{SQL} without a database connection.
}
